\name{JMH-package}
\alias{JMH-package}
\alias{JMH}
\docType{package}
\title{
\packageTitle{JMH}
}
\description{
\packageDescription{JMH}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JMH}
\packageIndices{JMH}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{JMH}

Maintainer: \packageMaintainer{JMH}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
