% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BrierJMMLSM.R
\name{print.BrierJMMLSM}
\alias{print.BrierJMMLSM}
\title{Print BrierJMMLSM}
\usage{
\method{print}{BrierJMMLSM}(x, ...)
}
\arguments{
\item{x}{x of class 'BrierJMMLSM'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print BrierJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
