% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.JMH.R
\name{print.JMH}
\alias{print.JMH}
\title{Print JMH}
\usage{
\method{print}{JMH}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'JMH'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of JMH object.
}
\seealso{
\code{\link{jmcs}}
}
\author{
Shanpeng Li
}
