% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PAJMMLSM.R
\name{summary.PAJMMLSM}
\alias{summary.PAJMMLSM}
\title{Print PAJMMLSM}
\usage{
\method{summary}{PAJMMLSM}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of class 'PAJMMLSM'.}

\item{digits}{number of decimal points to be rounded.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print PAJMMLSM
}
\seealso{
\code{\link{jmcs}, \link{survfitjmcs}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
