% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.AUCJMMLSM.R,
%   R/summary.ConcordanceJMMLSM.R, R/summary.JMMLSM.R, R/summary.MAEQJMMLSM.R,
%   R/summary.PEJMMLSM.R
\name{summary}
\alias{summary}
\alias{summary.AUCJMMLSM}
\alias{summary.ConcordanceJMMLSM}
\alias{summary.JMMLSM}
\alias{summary.MAEQJMMLSM}
\alias{summary.PEJMMLSM}
\title{Summaries of evaluation metrics for joint models}
\usage{
\method{summary}{AUCJMMLSM}(object, digits = 4, ...)

\method{summary}{ConcordanceJMMLSM}(object, digits = 4, ...)

\method{summary}{JMMLSM}(object, process = c("longitudinal", "survival"), digits = 4, ...)

\method{summary}{MAEQJMMLSM}(object, digits = 3, ...)

\method{summary}{PEJMMLSM}(object, error = c("MAE", "Brier"), ...)
}
\arguments{
\item{object}{object of class 'PEJMMLSM'.}

\item{digits}{number of decimal points to be rounded.}

\item{...}{Further arguments passed to or from other methods.}

\item{process}{for which model (i.e., longitudinal model or survival model) to extract the estimated coefficients.}

\item{error}{a character string that specifies the loss function.}
}
\value{
a list of matrices with conditional probabilities for subjects.

a list of matrices with conditional probabilities for subjects.

A table to summarize the model results.

a list of matrices with conditional probabilities for subjects.

a list of matrices with conditional probabilities for subjects.
}
\description{
Produce result summaries of a joint model fit.
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{JMMLSM}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}

\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
