% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitJMMLSMboot.R
\name{survfitJMMLSMboot}
\alias{survfitJMMLSMboot}
\title{Prediction in Joint Models}
\usage{
survfitJMMLSMboot(
  object,
  seed = 100,
  ynewdata = NULL,
  cnewdata = NULL,
  u = NULL,
  B = 100,
  method = c("Laplace", "GH"),
  quadpoint = NULL,
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{JMMLSM}.}

\item{seed}{a random seed number to proceed non-parametric bootstrap. Default is 100.}

\item{ynewdata}{a data frame that contains the longitudinal and covariate information for the subjects 
for which prediction of survival probabilities is required.}

\item{cnewdata}{a data frame that contains the survival and covariate information for the subjects 
for which prediction of survival probabilities is required.}

\item{u}{a numeric vector of times for which prediction survival probabilities are to be computed.}

\item{B}{the number of bootstrap samples to be generated. Default is 100.}

\item{method}{a character string specifying the type of probability approximation; if \code{Laplace}, then a first order estimator is computed.
If \code{GH}, then the standard Gauss-Hermite quadrature is used instead.}

\item{quadpoint}{number of quadrature points used for estimating conditional probabilities 
when \code{method = "GH"}. Default is NULL. If \code{method = "GH"}, then 15 is used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
This function computes the conditional probability of 
surviving later times than the last observed time for which a longitudinal 
measurement was available.
}
\seealso{
\code{\link{JMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
