% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.JMMLSM.R
\name{vcov}
\alias{vcov}
\alias{vcov.JMMLSM}
\title{Variance-covariance matrix of the estimated parameters for joint models}
\usage{
\method{vcov}{JMMLSM}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{JMMLSM}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a matrix of variance covariance of all parameter estimates.
}
\description{
Extract variance-covariance matrix for joint models.
}
\seealso{
\code{\link{JMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
